fix(linux/gtk4): DPI scaling and menu duplication fixes
- Implement proper DPI scaling using gdk_monitor_get_scale (GTK 4.14+)
  for fractional scaling support on Linux/GTK4
- Calculate PhysicalBounds correctly by multiplying logical coords by scale
- Fix menu items duplicating when creating new windows by adding
  processed flag to prevent re-processing menus
- Add safe type assertion helpers in screen example to prevent crashes
- Add CSS to prevent text selection during drag in screen example
- Document tiling WM limitations (Hyprland, Sway, i3) in official docs

+++ v3/pkg/application/webview_window_linux.go
@@ -107,7 +107,7 @@ func (w *linuxWebviewWindow) setMaximiseButtonEnabled(enabled bool) {
 
 func (w *linuxWebviewWindow) disableSizeConstraints() {
 	x, y, width, height, scaleFactor := w.getCurrentMonitorGeometry()
-	w.setMinMaxSize(x, y, width*scaleFactor, height*scaleFactor)
+	w.setMinMaxSize(x, y, int(float64(width)*scaleFactor), int(float64(height)*scaleFactor))
 }
 
 func (w *linuxWebviewWindow) unminimise() {
